feat: can pass -u flag to quant with file
This enhances the functionality of the `-u` or `--unfiltered-pl` flag.
Previously `-u` was simply a binary flag and could only be used with the
10xv2 or 10xv3 chemistry. Now, `-u` takes an *optional* path to an
external unfiltered permit list.  If the `-u` flag is passed with no
argument, then the behavior is the same as before (if 10xv2 or 10xv3,
the known permit list is downloaded if needed and then used). On the
other hand, if the `-u` flag is given an argument, that path is used
as the location of the unfiltered permit list. Note, if no argument
is provided and the chemistry is not 10xv2 or 10xv3, this is treated
as an error.

Grouped the parameters for the splici and direct reference sequences
to the `index` command logically under different help sub-headers.

With this commit, all of the limitation of `simpleaf` explicilty
listed in https://github.com/nf-core/scrnaseq/issues/93 have now
been addressed / lifted.  All of the things listed there are now
possible with the newest interface!