feat: added ability to set expected-ori in `quant`
This commit adds the ability to set the `--expected-ori` option
that is passed to the `generate-permit-list` command of `alevin-fry`.

This is an "optional" option.  If it is not provided explicitly, it
is set to "both" (allowing reads aligning in both orientations to pass
through), unless the chemistry is set as 10xv2 or 10xv3, in which
case it is set as "fw".  Regardless of the chemistry, if the user sets
this option explicitly, their choice is respected.