fix: parsing of multiple files for quant
Previously, the code to pass multiple files per read end in
the `quant` command did not work properly. The problem is that
even though the input is parsed as a `Vec<PathBuf>` in the
`simpleaf` argument parser, the clap derive parser
(the underlying argument parser we use) doesn't properly
treat the argument as a list unless `use_delimiter = true`
is set and unless `delimiter=','` is set. This commit
addressess this oversight, and multiple files in a
comma separated list can now be passed to the `-1` and `-2` options.
Thanks to @fmalmeida for uncovering this issue.