feat: quant from map and min-reads
This commit adds two new capabilities to
simpleaf:

- One can now run the `quant`
command giving a pre-existing mapping directory
that was previously generated by simpleaf, or
by salmon alevin, etc. In this case, quantification
will proceed from the RAD file in that directory and
will not require re-mapping the raw reads. In this
case the index and raw reads need not (in fact, can not)
be provided to the quant command. The argument parser
checks that *either* the --map-dir option is passed or
--index and -1 and -2 are passed.

- One can now pass `--min-reads` to the `quant`
command. The passed value will be propagated to the
alevin-fry `generate-permit-list` command to the parameter
of the same name.