feat: improve autodetection of index type
Previously, if the user passed `--use-piscem` to the
index command, they could avoid passing it to the quant
command. However, in that case, they had to always pass
the index directory (not the stem to the piscem index
files) and, more importantly, they *couldn't* also
pass the --use-piscem flag to quant. This created a
somewhat confusing usecase scenario.

This changes the behavior so that if the user
created the index with simpleaf and the `--use-piscem`
flag, then when they issue `quant`, they _can_ pass
`--use-piscem` (but they don't have to).  Further, if
the user passes an index file stem that looks like
it was created by `simpleaf index`, then `simpleaf`
will check if it was, and do the right thing in that
case.  Otherwise, it will proceed simply as directed
by the user (if `--use-piscem` was passed, assume that
the provided argument is a filename prefix stem, and
otherwise assume that is the path to a salmon index,
which is a directory).