feat: make piscem default, add --no-piscem flag
The piscem indexer and mapper is now the default.
The current command line parser will continue to accept
the --use-piscem flag, but this is the default and even
if no flag is passed, piscem will be used.

Additionally, there is a --no-piscem flag, which will trigger
the previous default behavior (using salmon-alevin) for indexing
and mapping.

Getting this to work (i.e. opposite boolean flags with one of
them being default), was kind of a pain and the existing solution
is kind of a hack (look in the comments in the clap derive
structs in simpleaf_commands.rs).

However, the current implementaion should have the desire behavior
and not introduce a breaking change by removing the `--use-piscem`
flag. However, that flag should be considered deprecated, and in
the future, there will likely be only the `--no-piscem` flag
instead.